Move RenderTemplateAsKind to internal/printutils
Decouples the template-to-YAML rendering from the cmd layer so it can be
reused. Signature takes kindName/apiVersion as primitives to keep
printutils free of pkg/schema (avoids a printutils → schema → utils →
resource → printutils cycle).
